+++ apps/web/CHANGELOG.md
@@ -1,5 +1,11 @@
 # web
 
+## 0.1.40
+
+### Patch Changes
+
+- improvements
+
 ## 0.1.39
 
 ### Patch Changes