+++ packages/components/card/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @jamsr-ui/card
 
+## 0.0.27
+
+### Patch Changes
+
+- improvements
+- Updated dependencies
+  - @jamsr-ui/progress@0.0.27
+  - @jamsr-ui/typography@0.0.27
+
 ## 0.0.26
 
 ### Patch Changes