+++ packages/components/dialog/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @jamsr-ui/dialog
 
+## 0.0.28
+
+### Patch Changes
+
+- improvements
+- Updated dependencies
+  - @jamsr-ui/button@0.0.27
+
 ## 0.0.27
 
 ### Patch Changes