+++ packages/components/drawer/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @jamsr-ui/drawer
 
+## 1.0.21
+
+### Patch Changes
+
+- improvements
+
 ## 1.0.20
 
 ### Patch Changes