+++ packages/components/header/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @jamsr-ui/header
 
+## 0.0.29
+
+### Patch Changes
+
+- improvements
+- Updated dependencies
+  - @jamsr-ui/divider@0.0.27
+
 ## 0.0.28
 
 ### Patch Changes