+++ packages/components/image-upload/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @jamsr-ui/image-upload
 
+## 0.0.27
+
+### Patch Changes
+
+- improvements
+- Updated dependencies
+  - @jamsr-ui/progress@0.0.27
+
 ## 0.0.26
 
 ### Patch Changes