+++ packages/components/link/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @jamsr-ui/link
 
+## 0.0.28
+
+### Patch Changes
+
+- improvements
+
 ## 0.0.27
 
 ### Patch Changes