+++ packages/components/popover/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @jamsr-ui/popover
 
+## 0.0.28
+
+### Patch Changes
+
+- improvements
+
 ## 0.0.27
 
 ### Patch Changes