+++ packages/components/textarea/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @jamsr-ui/input
 
+## 0.0.27
+
+### Patch Changes
+
+- improvements
+- Updated dependencies
+  - @jamsr-ui/button@0.0.27
+
 ## 0.0.26
 
 ### Patch Changes