+++ packages/config/tailwind-config/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @jamsr-ui/tailwind-config
 
+## 0.0.1
+
+### Patch Changes
+
+- improvements
+
 ## 0.0.17
 
 ### Patch Changes