+++ packages/config/typescript-config/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @jamsr-ui/typescript-config
 
+## 0.0.19
+
+### Patch Changes
+
+- improvements
+
 ## 0.0.18
 
 ### Patch Changes